requirements_complete.txt
# Core dependencies for Traffic Platform
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0

# Streaming & Processing
confluent-kafka>=1.9.0
pyspark>=3.3.0
delta-lake>=0.10.0
pyarrow>=10.0.0

# Data Warehouse
snowflake-connector-python>=3.0.0
dbt-snowflake>=1.3.0
dbt-core>=1.3.0

# Orchestration
apache-airflow>=2.3.4
apache-airflow-providers-apache-spark>=3.3.0
apache-airflow-providers-amazon>=5.0.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-redis>=3.0.0
apache-airflow-providers-celery>=3.0.0

# Cloud Services
boto3>=1.24.0
botocore>=1.27.0
awswrangler>=2.16.0

# Visualization & Dashboards
streamlit>=1.12.0
plotly>=5.10.0
seaborn>=0.11.0
matplotlib>=3.5.0
jupyter>=1.0.0
jupyterlab>=3.4.0

# Testing & Development
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-mock>=3.8.0
black>=22.6.0
flake8>=5.0.0
mypy>=0.971

# Security & Authentication
cryptography>=3.4.0
passlib>=1.7.0
python-jose>=3.3.0

# Configuration & Utilities
python-dotenv>=0.19.0
pydantic>=1.9.0
click>=8.1.0
rich>=12.5.0
typer>=0.6.0

# Database Connectors
psycopg2-binary>=2.9.0
redis>=4.3.0
sqlalchemy>=1.4.0

# API & Web Frameworks
fastapi>=0.78.0
uvicorn>=0.18.0
flask>=2.1.0

# Data Validation & Schema
jsonschema>=4.6.0
avro>=1.11.0
protobuf>=3.19.0

# Scheduling & Task Management
celery>=5.2.0
schedule>=1.1.0
apscheduler>=3.9.0

# Performance & Optimization
numba>=0.56.0
cython>=0.29.0
joblib>=1.1.0

# Logging & Debugging
loguru>=0.6.0
structlog>=22.1.0
sentry-sdk>=1.7.0

# Development Tools
pre-commit>=2.20.0
isort>=5.10.0
autopep8>=1.7.0
bandit>=1.7.0

# Documentation
sphinx>=5.1.0
sphinx-rtd-theme>=1.0.0
mkdocs>=1.3.0

# Machine Learning (Optional)
scikit-learn>=1.1.0
tensorflow>=2.9.0
torch>=1.12.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Time Series Analysis
prophet>=1.1.0
statsmodels>=0.13.0
pmdarima>=2.0.0

# Geographic & Location Data
geopandas>=0.11.0
shapely>=1.8.0
folium>=0.12.0

# Monitoring & Observability
prometheus-client>=0.15.0
statsd>=3.3.0
watchtower>=2.0.0

# Email & Notifications
smtplib>=0.2.1
email-validator>=1.2.0
slack-sdk>=3.15.0

# File Processing
openpyxl>=3.0.0
xlrd>=2.0.0
xlsxwriter>=3.0.0
csv>=1.0.0

# Compression & Archives
zstandard>=0.17.0
gzip>=1.0
zipfile>=0.1

# System & OS Utilities
psutil>=5.9.0
platform>=1.0.8
distro>=1.7.0

# Math & Statistics
scipy>=1.9.0
sympy>=1.10.0
statsmodels>=0.13.0

# Containerization
docker>=5.0.0
docker-compose>=1.29.0

# Code Quality
pylint>=2.14.0
mypy>=0.971
coverage>=6.4.0

# Build & Packaging
setuptools>=63.0.0
wheel>=0.37.0
build>=0.8.0
twine>=4.0.0

# Data Visualization Advanced
bokeh>=2.4.0
holoviews>=1.15.0
panel>=0.13.0
dash>=2.4.0

# Message Queues
pika>=1.3.0
kombu>=5.2.0
celery>=5.2.0

# Configuration Management
dynaconf>=3.1.0
configparser>=5.2.0
hydra-core>=1.2.0

# API Testing
httpx>=0.23.0
responses>=0.21.0
aioresponses>=0.7.0

# Excel Processing
openpyxl>=3.0.0
xlwings>=0.26.0
pandas>=1.5.0

# Natural Language Processing
spacy>=3.4.0
nltk>=3.7.0
transformers>=4.20.0

# Deep Learning Frameworks
tensorflow>=2.9.0
torch>=1.12.0
keras>=2.9.0

# Gradient Boosting
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0

# Clustering Algorithms
scikit-learn>=1.1.0
hdbscan>=0.8.0
umap-learn>=0.5.0

# Anomaly Detection
pyod>=1.0.0
scikit-learn>=1.1.0
isolation-forest>=0.1.0

# Time Series Forecasting
prophet>=1.1.0
statsmodels>=0.13.0
sktime>=0.11.0

# Model Monitoring
evidently>=0.2.0
whylogs>=1.1.0
arize>=1.0.0

# Data Validation
great-expectations>=0.15.0
pandera>=0.7.0
cerberus>=1.3.0

# Data Profiling
pandas-profiling>=3.2.0
sweetviz>=2.1.0
dtale>=1.22.0

# Data Lineage
openlineage>=0.16.0
marquez>=0.25.0
amundsen>=5.2.0

# Data Quality
great-expectations>=0.15.0
soda-core>=2.0.0
deequ>=1.2.0

# Data Ops
dbt-core>=1.3.0
great-expectations>=0.15.0
pytest>=7.1.0

# Data Monitoring
evidently>=0.2.0
whylogs>=1.1.0
arize>=1.0.0

# Data Pipelines
luigi>=3.1.0
prefect>=2.0.0
dagster>=0.14.0

# Data Streaming
faust>=1.10.0
bytewax>=0.16.0
quix>=0.4.0

# Data Lake
delta-lake>=0.10.0
iceberg>=0.2.0
hudi>=0.10.0

# Data Warehouse
snowflake-connector-python>=3.0.0
dbt-snowflake>=1.3.0
sqlalchemy>=1.4.0

# Data Mesh
fastapi>=0.78.0
graphql-core>=3.2.0
prisma>=4.0.0

# Data Product
streamlit>=1.12.0
dash>=2.4.0
plotly>=5.10.0

# Data Platform
kubernetes>=24.2.0
helm>=3.9.0
terraform>=1.2.0
